requirements.txt
openai==1.13.3
streamlit==1.32.0
langchain==0.1.11
langchain-openai==0.0.8
langchain-community==0.0.27
python-dotenv==1.0.1
duckduckgo-search==5.0
chromadb==0.4.24
pypdf==4.1.0
sentence-transformers==2.5
sqlalchemy==2.0.28